Add files via upload
Includes all the code required to run the program. It includes, in a modularized way, all the necessary functions: signin, login, questionnaire, matchmaking algorithm etc. The file includes a "menu" functions that allows you to run the program in the correct sequence: signin or login, questionnaire, editing of profile, viewing of matches, exit.